tech: Extracted async stack and pouch functions into its own file for readability and made fetchMore work with sorting
